arch/riscv: make RISC-V CSR accesses unsafe
Accessing RISC-V CSRs is an inherently unsafe and risky operation:
given they exercise large control over the state of the RISC-V hart,
they can be used to manipulate the current privilege mode, manage PMP
enforcement, etc.

Writing to CSRs can be used to violate Rust's memory safety
rules. This is especially true in the context of Tock, where we can
construct a simple example by looking at our system call handler: it
is sufficient for some untrusted code to store a non-zero
value (pointer to some normally inaccessible memory) into `mscratch`,
and a value to be written to that memory in a fixed register. The code
must then cause a trap, e.g., by issuing an illegal instruction. The
context switch code will incorrectly interpret this trap as a fault of
a process, and proceed to write the previous register file to a
"stack" indicated by `mscratch`. The kernel will continue execution
in the main loop, leaking the previous stack. The untrusted code has
successfully overwritten an arbitrary memory location.

The only reasonable solution is to mark all CSR accesses as
`unsafe`. While CSR reads are generally less problematic, even a CSR
read of a non-existant CSR can cause a trap, which would be yet
another way for some untrusted code to violate the kernel's
liveness (apart from looping & panicing). Because of this, and given
we provide access to the `ReadWriteRiscvCsr` types which inherently
allow writing to CSRs, this marks CSR register type construction and
CSR register type accesses through the `riscv::csr::CSR` struct as
unsafe operations.