kernel/platform/mpu: don't provide default impls for methods
Providing default implementations for the various `kernel::mpu::MPU`
methods makes it easy to forget a method in an implementation, and
would not cause compiler errors when adding a new method that should
be implemented. Default implementations should really only be used
when it is legal for an implemenation to not provide a specialized
method and fall back onto the default. This is not the case for any of
the `MPU` trait methods.

This instead moves those method implementations to the unit type,
which implements an `MPU` for boards that don't have one present.