Make StaticRef NonNull
This allows `Option<StaticRef<T>>` to be the same size as
`StaticRef<T>`. There is no usage of null `StaticRef` right now and it
would always be invalid: LLVM recognizes 0 as null and `.deref()`
constructing a null reference is immediate UB. Just in case, this adds a
check in svd2regs.py.