Add `capsules/tutorials` crate and add encryption oracle capsule
This adds `capsules/tutorials/`, a new crate for capsules used in Tock
tutorials. Capsules contained in this crate may not be particularly
useful, or even complete (e.g., they can be "checkpoint" versions for
following along in a tutorial), but they must compile. This ensures
that breaking changes in upstream Tock will also propagate to
kernel-code used in tutorials.

It further adds various checkpoints of the "encryption oracle"
capsule, to be built by participants in the upcoming Tock World 6
tutorial.

Co-authored-by: Amit Aryeh Levy <amit@amitlevy.com>