opentitan: fix aes_test test order
The AES tests are order-sensitive.

The CCM and GCM tests depend on the VirtualAES128 peripheral
configured during kernel initialization in `main.rs`.  In order
to operate correctly, the virtual peripheral depends on the
`set_client` calls to the low-level AES driver to configure the
virtual peripheral correctly.

The ECB, CBC and CTR tess, on the other hand, call `set_client`
on the AES driver, thus breaking the VirtualAES128 configuration.

Run the CCM and GCM tests first, then run the ECB, CBC and CTR tests.

Signed-off-by: Chris Frantz <cfrantz@google.com>