tickv: remove key search function
This isn't ideal, but currently tickv does not correctly implement
search and this fixes a bug when keys hash to region 0 in the store.

Currently, if there is a key stored in region 0 and a get() for a key
hashes to region 0 tickv will enter an infinite loop checking the same
flash region for the key.

I suppose it would be better to correctly implement the search, but it
is difficult to do correctly.