libraries: tickv: Fixup the broken region search
Fixup the broken region search as decsribed at:
https://github.com/tock/tock/pull/3617

We already have an existing test that accesses multiple regions, so we
also need to update that test now that we are correctly searching
regions.

As part of this change we convert the region from an isize to a uszie as
well.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>