chips/earlgrey: convert CONFIG to trait, make board-configurable
The EarlGrey chip exists in multiple variants, such as the CW310 FPGA
instantiation, the Verilator simulation, or eventually the taped out
chip. While the "silicon" will be idential in those chips, they can
still have slightly different configurations for attributes such as
clock speeds.

To drive these configuration flags, instead of using Cargo feature
flags and passing them down to the `earlgrey` chip crate, this changes
the chip crate to accept a generic `EarlGreyConfig` argument wherever
necessary. This is a trait containing a number of constants. By
implementing this trait on variant-less enums, it is guaranteed to not
generate any runtime code.

This change is in the spirit of reducing Tock's reliance on
conditional compilation: by plugging in a type-parameter implementing
a given trait, all code is always type-checked to adhere to this
trait's constraints, regardless of the actual configuration used
during the build. Conditional compilation is solely used to switch
between two type-aliases which expose implementations of the
aforementioned trait.

In practical terms, this allows downstream users to build a customized
board crate (where the board logically is also responsible for driving
clocks, etc.), without maintaining patches for the chip crate. The
chip crate is reduced to code related to the actual silicon.