virtual_alarm: Do not disarm when setting client.
When using the Component::finalize paradigm, it is useful to only put
static initialization and client connections in the finalize method. All
business logic should move outside of the finalize method. This makes
the business logic easier to unit test.

It is possible that the business logic needs to start an alarm when the
object is created (e.g. to debounce some value). The constructor of the
actual capsule is a good place to put the business logic if that capsule
must always run the business logic when the capsule is created.

Here lies the issue: if the constructor of the capsule starts an alarm,
there is no client yet, but the Component::finalize method will hook up
the alarm_client relationship after the capsule object has been created.
If the set_alarm_client method disarms the alarm, then the alarm that
was set in the constructor is canceled.

Wait, isn't there a race condition or something weird going to happen if
one sets an alarm before setting a client? Generally no, since TockOS is
single threaded and all the alarm callbacks are meant to be handled in
the main kernel loop.

An alternative would be for the capsule to create an init method that
must always be called by the Component::finalize after the finalize
method hooks up all of the client callback relationship. This
allows a capsule object to be created in a half initialize state, which
developer may want to prevent.

Lastly, the virtual_timer doesn't disarm anything when updating the
client; only virtual_alarm does. Let's make things more consistent as
well.

Change-Id: I112c11b9b1b800df61098e8afd7241817beef049