TicKV: Zeroizing a key to remove data
In the documentation of TicKV:

    This is done because changing a 1 to a 0 in flash can be done with
    a write to a single byte (where only 1 bit changes). While changing
    a 0 to a 1 requires an erase of the entire region.

This change will Zeroize a key without erasing by utilizing the
ability to change a bit from a 1 to a 0 through a write as used in the
invalidate_key function. This logic is applied throughout the value of
the key without modifying the header. This is done to improve security
without degrading the flash device with erase. Additionally, the valid
bit will be flipped, allowing for the garbage collector to function as
normal.