libraries/tickv: Implement key Zeroise
PR https://github.com/tock/tock/pull/3715 originally attempted to
implement a key zeroise function. Unfortunatley it was merged without
proper reivew or testing and as such it doesn't zeroise the data and the
documentation doesn't cover a few issues with the implementation
as well.

This PR updates the logic to correctly zeroise keys and checksums. It
also updates the tests to check that this has correctly occured. We also
update the documentation to clearly describe what we do and don't
zeroise and that we don't do multiple passes or other fancy data
destructure mechanisms, so we rely on the flash hardware.

While where we update the spelling to match the other spelling in TicKV
and the wikipedia article.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>