boards: particle_boron: add start() support
Adds a start function to setup the board. This function can never be
inlined ensuring that it's stack frame is popped upon return. Before,
the stack space associated with static_inits would be wasted. With this
change, we can use a smaller stack.

This fixes the stack overflow caused by the `kernel` command in the
pconsole. The alternate solution is to increase the stack_size. But this
is more elegant :)

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>