lowrisc: usb: change while to if
This while loop can only ever execute once because there are break
statements in all of the valid match arms.

This was found via a new clippy lint.

This patch makes the code match what is actually happening. I don't know
if it should be a loop and the breaks are the issue, however.