nrf52: i2c: add event/irq clearing
Bug: When switching between master and slave mode, it is possible
that there are unserviced pending events set (harmless). For example, from master
mode after a write, `txstarted` maybe set. When we do a
`handle_interrupt()` thereafter in slave mode for a receive, we will fall
through into the `txstarted` branch as it was not cleared.

This change ensures we clear such a state when ping-ponging between
master/slave modes.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>