tock-registers: added `debug_registers` feat to generate better `Debug`
So, the idea is to get an easy view of the data inside, especially for
more complex fields.

Currently it does the following:
- It will try to parse the enum if it can, otherwise it will just print
  the value as a number (not hex).
- If the value is not enum (wasn't created with [<fields>]), it will
  just print the number.
- Unless its 1 bit, it will print it as a bool.

This change is completely hidden from the API, and shouldn't change
anything, the feature is not enabled by default.
When the feature is disabled, the resulting code is exactly the same
before this commit.