doc: remove courses
These courses are very out of date and have little (if any) learning
value for current tock. They are all for v1, and have not been updated
in at least three years. At this point they are more misleading than
helpful and the guides in the Tock Book are a much better place to look.