doc: move main docs to book
The book has better formatting, search, discoverability. This is a start
to move our documentation to one central place.

However, documentation about code development and this repository makes
sense to keep here. So documentation that is less about Tock as an
operating system and more Tock as a software engineering exercise is
kept here.

This is still a start so I think there is more to be done to figure out
exactly where we want things.