kernel/platform/mpu: remove MPU::clear method
The MPU trait's `clear` method is never used in Tock and has unclear
semantics. To ensure that the kernel can access all userspace memory
regions, it can call the `disable_app_mpu` method. After this method,
the MPU may still hold entries applicable to kernel-mode
code. However, on some platforms (e.g., ePMP), these regions cannot be
removed. With the `clear` method it is implementation defined which
regions are removed and which are not. It is further unclear how this
method interacts with other trait methods (e.g., `disable_app_mpu`,
then `clear`, then `enable_app_mpu`). Thus it is not a particularly
useful method for generic kernel code to rely on, and further should
not be required.