Avoid creating intermediate Rust references for taking pointers
Avoid creating intermediate references for taking pointers to
variables or extern statics during process loading, setting up the
RISC-V PMP, creating non-volatile storage drivers, and when taking the
addresses of registers in chips. Extern statics are inaccessible in
safe Rust code by default. However, when we perform an operation such
as `&_sapp as *const u8`, we create a Rust reference to the underlying
memory that is (a) safely dereferencable, and (b) does not necessarily
conform to Rust requirements concerning initialized memory.

This PR switches those occurrences to use `core::ptr::addr_of` (and
its mutable sibling), a macro explicitly designed to return the
address of some memory location without creating an intermediate
reference.

Suggested-by: Alyssa Haroldsen <kupiakos@google.com>