arch/cortex-m0: hard_fault_handler: fix incorrect return to PSP stack
This fixes a critcial bug in the ARMv6m hard-fault handler. Prior to
this fix, it is possible (depending on the compiler) that the
hard-fault handler may return to an application's context (the PSP
stack), but *in privileged handler mode*. This effectively and
trivially allows an application to execute arbitrary code at kernel
privileges.

It mirrors changes of ad2acc8190 ("arch/cortex-m: hard_fault_handler:
fix incorrect return to PSP stack"). See that commit for more
information on the issue.