arch/cortex-m: hard_fault_handler: fix incorrect return to PSP stack
This fixes a critcial bug in the ARMv7m hard-fault handler. Prior to
this fix, it is possible (depending on the compiler) that the
hard-fault handler may return to an application's context (the PSP
stack), but *in privileged handler mode*. This effectively and
trivially allows an application to execute arbitrary code at kernel
privileges.

Tock's Cortex-M hard-fault handlers are split into a base
handler (implemented through a `#[naked]` function that does not
generate a function prologue / epilogue), and a non-naked `_continued`
function which can contain arbitrary Rust code. Once the base handler
has done sufficient preparations (make stack space for the subsequent
handler, in case of a kernel stack overflow), the handler jumps to the
`_continued` function. This function is expected to only return if the
hard-fault has been caused by unprivileged user-mode code. In this
case, the CONTROL register is set to switch to privileged thread mode,
and the link register (LR) is set to `0xFFFFFFF9` (return to Thread
mode, restoring state from the main stack pointer MSP). Finally, upon
return, the naked handler is supposed to branch to this LR address:

    00020e4c <_RNvCsiXETrHxse8y_7cortexm26hard_fault_handler_arm_v7m>:
       20e4c:       f04f 0100       mov.w   r1, #0
       20e50:       f01e 0f04       tst.w   lr, #4
       [...]
       20e78:       46a5            moveq   sp, r4
       20e7a:       f000 b807       b.w     20e8c <_RNvCsiXETrHxse8y_7cortexm36hard_fault_handler_arm_v7m_continued>
       20e7e:       4770            bx      lr

However, because the `_continued` function is a non-naked function,
the Rust compiler / LLVM may generate an arbitrary function
epilogue. This epilogue may overwrite the LR register upon return.

    unsafe extern "C" fn hard_fault_handler_arm_v7m_continued(
        faulting_stack: *mut u32,
        kernel_stack: u32,
        stack_overflow: u32,
    ) {
        use core::arch::asm;
        if kernel_stack != 0 {
            [...] // This branch never returns
        } else {
            asm!(
                [...]
                "movw LR, #0xFFF9",
                "movt LR, #0xFFFF",
                options(nostack),
             );
        }
    }

The above code can, for instance, generate the following assembly:

    00020e8c <_RNvCsiXETrHxse8y_7cortexm36hard_fault_handler_arm_v7m_continued>:
       20e8c:       b5fe            push    {r1, r2, r3, r4, r5, r6, r7, lr}
       20e8e:       af06            add     r7, sp, #24
       [...]
       20ebc:       f3bf 8f6f       isb     sy
       20ec0:       f64f 7ef9       movw    lr, #65529      @ 0xfff9
       20ec4:       f6cf 7eff       movt    lr, #65535      @ 0xffff
       20ec8:       b006            add     sp, #24
       20eca:       bd80            pop     {r7, pc}
       [...]

In this case, the `pop {r7, pc}` instruction returns from the
function, but to the stacked `lr` instead of the newly written return
address. This stacked `lr` value is the faulting address in case of an
instruction access fault by userspace. Thus, userspace can execute
arbitrary code in the privileged handler mode.

This commit fixes this issue by calling the `_continued` function with
a proper branch and link instruction, and unconditionally jumping to
the `EXC_RETURN` code upon function return.