opentitan: Rearrange UARTs and I2Cs.
OpenTitan has a convention for which UARTs and I2C peripherals are mapped to which pins. This PR updates the pinmux layout to match that convention.

With this change, both the RX and TX pins of UART0 are mapped to the SAM3X chip on the cw310, allowing us to interact with the Tock console over USB without extra hardware.