chips/apollo3: iom: I2C: Split handing of THR and CMDCMP interrupts
Previously if a THR (threshold) or CMDCMP (command complete) interrupt
occured we handled them both the same way. This works for small amounts
of data as we could finish sending all the data before triggering the
command complete callback.

If trying to send large payloads (128 bytes for example) we actually get
a number of THR interrupts before we have completed the write operation.
The existing logic would return complete before we had finished the
operation.

This patch splits the hadling of THR and CMDCMP so we will only return
if the hardware provides a CMDCMP interrupt.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>