stm32f4xx: handle usart overrun error
Similarly to the stm32f303xc usart driver, the overrun error handler
clears the error flag, sets the rx_status back to Idle and notifies the
client. In this case, the ongoing DMA transfer is also aborted.
Furthermore, the tx complete handler is now executed only if the TC flag is
actually set, to avoid running it upon an error interrupt event.

This handler avoids the usart driver to hang indefinitely after an overrun
error occurs.