Alter Frame struct to support preformed packets
These changes alter the Frame struct to support forming a frame object
from a buffer that is provided from userland preformed (i.e. there is
not a need to form the header).

For a preformed object, the frame info is irrelevant. Wrapping this in
an option allows this to be ignored in the case of when a provided
packet is preformed.