capsules: i2c_master: Use I2CMux
To allow an I2C bus to have devices created by the board and still be
exposed to userspace let's convert the I2CMasterDriver to use an
I2CDevice instead of a I2C hardware.

This allows muxing an I2C bus with both userspace and devices in the
kernel.

This commit also adds a I2CMasterDriverComponent to help in creating the
I2CMasterDriver.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>