process_console: fix wrong byte removal in Delete EscKey handler
The Delete escape key handler was removing from the current cmd buffer
the character *before* the cursor instead of the character *under* the
cursor as it should. This led to a panic upon deletion of the first
character (cursor == 0) in the current cmd due to an 'attempt to
subtract with overflow'.