process_console: fix char deletion on history commands
Deleting any character after navigating the history through the Up or
Down key escapes, could result in a panic due to an 'attempt to subtract
with overflow'. This because the BS and Delete handlers were trying to
call `delete_byte` on the 'unfinished' cmd buffer which could even be
empty, since it was not being updated with a copy of the cmd from
history.
With this fix, after history is accessed and a character deletion is
requested, the modified cmd from history is copied into the 'unfinished'
cmd buffer instead of erroneously deleting a character from a potential
empty buffer. This reflects the case when the history is accessed and a
character insertion is requested. Also, the call to `change_from_cmd` is
now unnecessary since the 'unfinished' buffer gets updated just after a
character deletion.