License checker: Use the fallback syntax for `.txt` files.
Syntect was loading a "plain text" syntax for files with the `.txt` extension. This results in a "license header missing" error, because syntect doesn't parse any of the contents as comments. This change makes the license checker use the fallback parser instead, which is designed to handle plain text files (and files of unknown type) well.

Fixes issue #3880.