capsules/alarm: rework to use A::Ticks internally, simplify logic, add tests
Before this change, and in particular with the introduction of #3973,
it is very hard to reason about the wrapping logic in the alarm driver
with > 32 bit timers. This is because the alarm driver used to deal
only with 32-bit values internally. However, it would not record
whether those values were potentially stemming from a timer of a
smaller timer. This then causes issues when it will set a new alarm in
its wrapping logic.

Apart from that, the fact that it first converted to 32 bit, and then
back again to the underlying timer width, up- and down-scaling as
necessary, is very confusing. The code interleaves iterating over
timers, computing wrapping, and controlling the hardware in a way that
makes it hard to reason about, and virtually impossible to unit-test.

I refactored the driver to clearly separate out this logic:

- `Expiration` now always refers to a set alarm. `Option<Expiration>`
  refers to an alarm that can be disabled. Distinguishing between
  these on a type level (versus one single `enum Expiration` for both)
  allows us to get rid of a many checks.

- `earliest_alarm` iterates over a list of alarms, invoking a callback
  on each expired one, and returns the earliest non-expired one. This
  is very easily unit-testable and allows us to combine callback- and
  re-arm operations into a single handler.

- `process_rearm_or_callback` is this handler. It is a thin wrapper
  around the above, and controls the hardware (this part is not easily
  testable).

- `rearm_u32_left_justified_expiration` does all the magic wrapping /
  shifting logic. It is not clobbered with any details about the
  underlying alarm, but instead just converts a 32-bit left-justified
  ticks value into any underlying A::Ticks, regardless of width. It
  does appropriate wrapping for smaller and larger timers. This is
  extensively unit-tested.

The logic in the `Driver` implementation is simplified as well.