kernel/collections/queue: rename `remove_first` to `remove_first_matching`
`Queue`'s `dequeue_specific` was renamed to `remove_first` in
ab293c54c36a ("Rename dequeue_specific to remove_first").

While I don't disagree with the fact that `dequeue_specific` is an
oxymoron and a bad name, arguably `remove_first` carries its own
baggage. We're talking about collections where "first" has meaning,
often used interchangably with "head". Looking over the code in #3577
I was quite surprised with the complexity of this method's
implementation in `RingBuffer`, for what I thought would just be a
`dequeue` operation---removing the _first_ / _head_ element.

This is an effort to clear up that confusion. This is not a hill worth
dying on for me, so if anyone has strong feelings otherwise, feel free
to close this.