chips: nrf52: use only spi0 and twi1
SPI0 and TWI0 are the same peripheral, and SPI1 and TWI1 are the same
peripheral. Boards cannot use both 0 or both 1 at the same time.

The current default peripherals expose both, allowing boards to use two
SPI or two TWI, and they must be careful to not use both 0 or both 1
peripherals.

This changes the default to be more opinionated and only expose one spi
and one twi. This avoids the problem entirely for boards that use the
default peripherals. We can of course support other peripheral groups
for users who need say 3 SPIs or 2 I2C peripherals.