tickv: Remove the offset from read_region()
read_region() originally had a `offset` argument. The idea was to avoid
reading an entire flash region and only read part of the region, by
offsetting into the region.

This was never actually supported, the idea was to be able to support it
without API breakages, but in practice I don't think that will be
possible as we will need some API changes no matter what.

As offset is never used let's just remove it. This will simplify things
a little bit and it's easy to re-add in the future if we want to support
partial page reads.

This also fixes https://github.com/tock/tock/issues/4066

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>