chips: apollo3: stimer: Ensure alarm occurs if set in the past
If a userspace application calls libtocksync_alarm_delay_ms(0) we
previously would not generate an interrupt as we would set the compare
time to the current tick count. That means we have to wait for the ticks
to wrap all the way around to generate an interrupt.

Instead if the timer should have already expired let's just manually
set the interrupt and return.

Signed-off-by: Alistair Francis <alistair@alistair23.me>