kernel: process checking: wrap type & usize
Rather than just us a `usize` directly, this wraps the usize in a struct
so we can give it a name and documentation.

This also changes the nonzero usize to a normal usize. Using a nonzero
usize is a nice trick to make the option<usize> smaller, but not being
able to use zero is a bit of usability hit. I don't think saving a
couple bytes is worth it.