boards: linker: fix root cause of relocation bug
Currently we rely on `-nmagic` for the linker not to do funny things
with the beginning of the relocation section such that it always
actually is stored starting at `_etext`. This is a hack. Instead, we can
force the load-memory-address (LMA) of the relocate section to be
exactly `_etext` using `AT (_etext)`.

see <https://sourceware.org/binutils/docs/ld/Output-Section-LMA.html>