chips: apollo3: Support the flash controller
The Apollo3 flash controller is not documented in the datasheet and not
even documented in the vendor HAL. Instead the vendor HAL calls to
special ROM functions to do flash read, write and erase operations. This
is to allow operating on flash while also executing from it.

This commit adds support for the flash by calling the same ROM
functions. There doesn't appear to be any other way to write or erase
flash.

Signed-off-by: Alistair Francis <alistair@alistair23.me>