segger_rtt: make sound + move to own chip crate
Fixes an issue in the segger rtt capsule where the capsule implicitly
exposes memory to the SWD host while retaining and modifying a `mut`
(i.e. unique) reference to it internally. The "solution" is to do this
explicitly with shared volatile cells and unsafe, and to move the
capsule to its own chip crate.

Part of the reasoning is that this is actually quite similar to DMA, in
the sense that we need to control access to otherwise arbirary memory
accesses from a "peripheral" (it's just that the peripheral is a
debugger).