segger_rtt: implement uart Receiver
Implements the receive path for Segger RTT. Unfortunately, RTT has no
mechanism of notifying (e.g. via an interrupt) for the availability of
RX data so, instead, when the waiting for outstanding data, the
implementation just polls the RTT buffer every 100 milliseconds---fast
enough to be unnoticable to a human but infrequent enough to interrupt
most computations.