Fix ProcessArray
Current pattern for process array is UB as both the scheduler and process loading logic have concurrent &mut references to it.

This change adds Cells internally. It also adds in new wrapper types / constructor methods so that boards do not need to be so aware of the exact types they are allocating. This makes future changes to the main process array not need to touch every board (like this has to).

Change-Id: I8f200e0051af618076c16b0d512bf20863342769