MetaPtr + usize/u32 changes
This changes some u32s into usizes as appropriate.

It also introduces a new MetaPtr type that is a pointer that has explicit target dependant metadata. On many platforms, this will just be a wrapper around usize.
On CHERI platforms MetaPtr will be a capability.
See later commit.

It also adds a new syscall encoding `encode_syscall_return_metaptr`
That is intended to work for both 32/64 platforms with or without this extra metadata.

Change-Id: I40faa11c1fd53debc6e9b21d00772660cacf8cab