ot/earlgrey-cw310: support alternative linker script for tests
This commit fixes the hack where OT currently relies on overwriting the
linker script in the target directory, instead using a custom build.rs
to allow this board to override the linker script in use based on the
value of an environment variable. This PR also includes
some cleanup to remove some lines that I believe are no longer necessary
in the current build system. These line were scattered around the "test"
rule for various boards and would directly copy files into the build
directory, but I that is already handled by Cargo today.