move to region-level magic headers
Instead of having one single magic header value at the start of
userpsace, this commit changes it so there is a magic value at the start
of every region header. This magic value is set to the XOR of the
region owner's ShortID, region length, and a constant to distinguish
what version of the capsule the header was allocated from.

This has a few benefits:
- As this capsule is modified, it can support previous version of region
  headers.
- Previously, the capsule looked for a special ShortID value of 0 to
  determine if the region traversal was done. Now, it just checks if
  the magic value is invalid (not equal to an XOR between the region ShortID,
  length, and any of the supported version numbers). When an invalid
  header is reached, the traversal can be terminated and we can start
  allocating new regions there.