lpm: restrict updates to in-frame rows
Both the transfer and screen refresh on the LPM are slow, but it's
possible to transfer only the relevant rows for the current frame (the
whole row needs to be transfered, not just the relevant columns). This
dramatically improves update latency for applications that update only a
portion of the screen (e.g. update some text) from unusable to basically
seemless.