kernel/syscall: split out encode_syscall_return, create TRD104 subset
Because all architecture crates depend on the kernel crate, we use it
not just for architecture-agnostic core kernel infrastructure, but
also architecture-specific code that happens to be shared between two
or more `arch` crates. Prior to this change, we conflated these
helpers with the core kernel infrastructure in `syscall.rs`.

With this change, we split out the 32-bit and TRD 104-specific
`encode_syscall_return` helper, shared between the 32-bit RISC-V and
Cortex-M architecture implementations, into its own dedicated module
under utilities.

We create a separate `SyscallReturn` subset for those return values as
specified in TRD 104, and create a function that translates between
these types. This allows architectures which use these functions to
guarantee that they are conformant to TRD 104 by first explicitly
converting into this type, and then encoding it into their stored
registers using the TRD 104 encoding helper.