kernel: split out encode_syscall_return to utilities module
Because all architecture crates depend on the kernel crate, we use it
not just for architecture-agnostic core kernel infrastructure, but
also architecture-specific code that happens to be shared between two
or more `arch` crates. Prior to this change, we conflated these
helpers with the core kernel infrastructure in `syscall.rs`.

With this change, we split out the 32-bit and TRD 104-specific
`encode_syscall_return` helper, shared between the 32-bit RISC-V and
Cortex-M architecture implementations, into its own dedicated module
under utilities.

We also define a `TRD104SyscallReturn` enum type that defines the
subset of `SyscallReturn` values defined in TRD104, as well as a
conversion method between those types. This assures, on a type-level,
that changes to the core-kernel's `SyscallReturn` enum will not
propagate to userspace, except for an explicit conversion method.