kernel: make {Readable,Writeable}ProcessBuffer unsafe traits
As raised in https://github.com/tock/tock/issues/3757, users of these
traits will need to rely on the implementations of `ptr` or the newly
added `mut_ptr` to be correct, if they want to be able to dereference
the returned pointer.

By making these unsafe traits, we force implementors of these traits
to guarantee that their method implementations comply with the
semantics described in the documentation of `ptr` and `mut_ptr`
comments. This makes it possible for these APIs to be used in a sound
fashion, without risking that a safe implementation will not adhere to
this contract and potentially return invalid pointers (other than in
the zero-length case described).