chips: apollo3: iom: Work with new SPI HIL
The new SPI HIL changes how the length is supplied, so we need to update
to handle that correclty. At the same time we can also use get() to
access the buffers to remove panics.

We also sneak in yet another errata delay, as that improves the
reliability of getting the correct FIFO sizes.

Signed-off-by: Alistair Francis <alistair@alistair23.me>