arch: cortex-m: v7m: check SPSEL in EXC_RETURN
In the exception handling assembly for the cortex-m platform we have
been using the full value of `EXC_RETURN` for checking if we are coming
from userspace and for directing where we should go to after the
exception handler returns. This has been effective, but has two
drawbacks:

1. It will not work for full floating point-enabled hardware/software
   that uses the larger exception frame.
2. It will not work for the v8m architecture which uses more of the bits
   that are reserved on v7m.

As we look to support more cortex-m architectures, this change will make
that easier. This should have no effect on our current v7m no fp
platforms.