pconsole: `boot` cmd use process.start()
With the changes to the process loading, `process.try_restart` only has
an effect if the process is running. If the process is terminated it
does nothing. The function we need to call is `process.start()` for the
boot command.

With this change we can actually start a terminated process using the
process console.