boards: make: update stable compiler if needed
Apparently there is no way to specify a minimum rust version for a crate
and then have rustup just install a suitable version for you
automatically. Why? I have no idea. So, we implement this feature in our
makefile.

This checks if the board is using stable, and if so, checks if there is
an error from the stable compiler being too old. If so, it waits 5s and
then installs a newer toolchain automatically.