chips: apollo3: Disable UART when not in use
The UART uses a relativly large amount of power just being enabled. So
instead of enabling it at startup let's instead only enable UART when
required and then disable it afterwards.

As part of this we want to ensure that interupts are only handled for
the TX/RX FIFO events.

Signed-off-by: Alistair Francis <alistair@alistair23.me>