doc: trd104: fix pointer->parameter
I believe that referring to the "appdata" in subscribe as an
"application data pointer" is unintentional, and that it should be
"application data parameter" to match the earlier description. I believe
this stems from userspace largely using that parameter to store a
pointer, but we don't require that (and can't enforce it). This
clarifies the document.