kernel: remove debug_enqueue
This debugging feature never caught on. It was added in 04/2020 and
there are no upstream boards that are using it.

Since then we now have `capsules/core/src/console_ordered.rs` which
_may_ help in some use cases where ordering was desired.

Also, boards can use a larger buffer to our existing debug
infrastructure if they need more debugging output.

In line with
https://book.tockos.org/doc/design#merge-aggressively-archive-unabashedly
this patch removes the feature.