arch/rv32i/pmp: restrict physical address space on RV64 to 56 bit
RV64 platforms support only a 56 bit physical address space. For this reason
(and because addresses in `pmpaddrX` CSRs are left-shifted by 2 bit) the
uppermost 10 bits of a `pmpaddrX` CSR are defined as WARL-0.

To avoid unexpected behavior when configuring the PMP with such
violating addresses, this commit checks that any `pmpaddrX` values
represented and encoded by TORRegionSpec or NAPOTRegionSpec instances
maintain the above invariant.