arch/rv32i: allow full address range to be covered by {TOR,NAPOT}RegionSpec
This changes the TORRegionSpec and NAPOTRegionSpec struct's internal
representation to store pre-generated `pmpaddrX` CSR values, which
removes artificial restrictions to them covering only addresses in
`[0; usize::MAX)` (that is, excluding address 2**XLEN - 1) due to the
prior internal representation using an (addr, len) specification.

It retains the constructors---albeit renamed to clarify their
semantics---to generate such a pmpaddr value from a size and
address. It also adds a constructor to build a TORRegionSpec and
NAPOTRegionSpec directly from a (pair of) raw pmpaddrX CSR
values. These constructors still uphold the same invariants over these
types, but need to perform less checks due to the nature of the
`pmpaddrX` CSR encoding.

It removes the `addr` and `len` methods that these types previously
had, as they now can return values out of range for the `usize`
integer type. Convenience-methods to introspect the `RegionSpec`
structs apart from retrieving the `pmpaddrX` CSR value may be re-added
later, either with using a 64-bit integer type (which can hold all
legal addresses and lengths even for 64-bit platforms), or alternative
representations. This is not currently required.

The change also adds an additional convenience method to create a
`NAPOTRegionSpec` from a pair of start and end addresses, removing
pointer arithmetic in the board's initialization routines.