segger_rtt: isolate unsafe components
This ended up slightly more substantial than anticipated. The goal
was to split the segger_rtt implementation into separate safe and
unsafe parts.

In practice, the unsafe parts of RTT are contained to interfacing
with the memory behind the up/down channels. Separating that out
cleanly required refactoring to pull the channel access logic out
of the main RTT implementation and into dedicated implementations
for the up and down channels themselves.